[OPIK-6545][FE] Refactor upload format detection and human-labeling logic into shared `file` utilities
- Moved `detectUploadFormat` and `formatToHumanLabel` functions from individual components to the shared `file` module.
- Replaced duplicated inline logic in Add/Edit Dataset dialog components with shared utilities.
- Improved maintainability by centralizing upload format-related functionality.
